fix tracing test imports
