+++ test/moonwall.config.json
@@ -421,7 +421,7 @@
       ],
       "multiThreads": 4,
       "envVars": ["DEBUG_COLORS=1"],
-      "reporters": ["html", "basic"],
+      "reporters": ["html", "default"],
       "foundation": {
         "type": "dev",
         "launchSpec": [
