+++ test/suites/tracing-tests/test-trace-1.ts
@@ -1,70 +1,8 @@
-import {
-  beforeAll,
-  customDevRpcRequest,
-  describeSuite,
-  expect,
-  DevModeContext,
-  deployCreateCompiledContract,
-} from "@moonwall/cli";
-import { alith, ALITH_PRIVATE_KEY, createEthersTransaction } from "@moonwall/util";
-import { Abi, encodeFunctionData } from "viem";
+import { beforeAll, customDevRpcRequest, describeSuite, expect } from "@moonwall/cli";
+import { ALITH_PRIVATE_KEY, alith, createEthersTransaction } from "@moonwall/util";
+import { encodeFunctionData } from "viem";
+import { nestedSingle } from "../../helpers";
 import BS_TRACER from "../../helpers/tracer/blockscout_tracer.min.json" assert { type: "json" };
-
-export async function createContracts(context: DevModeContext) {
-  let nonce = await context.viem().getTransactionCount({ address: alith.address as `0x${string}` });
-  const { contractAddress: callee, abi: abiCallee } = await deployCreateCompiledContract(
-    context,
-    "TraceCallee",
-    { nonce: nonce++ }
-  );
-
-  const { contractAddress: caller, abi: abiCaller } = await deployCreateCompiledContract(
-    context,
-    "TraceCaller",
-    { nonce: nonce++ }
-  );
-  await context.createBlock();
-
-  return {
-    abiCallee,
-    abiCaller,
-    calleeAddr: callee,
-    callerAddr: caller,
-    nonce: nonce,
-  };
-}
-
-export async function nestedCall(
-  context: DevModeContext,
-  callerAddr: string,
-  calleeAddr: string,
-  abiCaller: Abi,
-  nonce: number
-) {
-  const callTx = await createEthersTransaction(context, {
-    to: callerAddr,
-    data: encodeFunctionData({
-      abi: abiCaller,
-      functionName: "someAction",
-      args: [calleeAddr, 6],
-    }),
-    nonce: nonce,
-    gasLimit: "0x100000",
-    value: "0x00",
-  });
-  return await customDevRpcRequest("eth_sendRawTransaction", [callTx]);
-}
-
-export async function nestedSingle(context: DevModeContext) {
-  const contracts = await createContracts(context);
-  return await nestedCall(
-    context,
-    contracts.callerAddr,
-    contracts.calleeAddr,
-    contracts.abiCaller,
-    contracts.nonce
-  );
-}
 
 describeSuite({
   id: "D3601",
